requirements.txt
Flask==3.0.1
pillow==10.2.0
numpy==1.26.3
tensorflow==2.15.0